ret type of adc functions

+++ main.c
@@ -35,12 +35,12 @@
 #define PWM_CK_DIV_BITS _BV(CS02) | _BV(CS00)
 #endif
 
-static uint8_t adc_pot_read()
+static uint16_t adc_pot_read()
 {
         return 0;
 }
 
-static uint8_t adc_feedback_read()
+static uint16_t adc_feedback_read()
 {
         return 0;
 }
